backend/requirements.txt
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pydantic-settings
flask
flask-cors
flask-sqlalchemy
flask-migrate
flask-login
flask-wtf
flask-bcrypt
alembic
google-generativeai
google-auth-oauthlib
google-auth-httplib2
aiofiles
httpx
aiosqlite
openai
slack-sdk
slack-bolt
reportlab
markdown